update to move profile stepper and update to square paper
