+++ src/components/App.js
@@ -12,7 +12,7 @@ import Footer from '../components/Footer';
 import Home from '../components/Home';
 import Login from '../components/Login';
 import Register from '../components/Register';
-import ProfileStepper from '../components/ProfileStepper';
+import ProfileStepper from './Profile/ProfileStepper';
 import ProfileMainView from '../components/Profile/MainView';
 import FinanceMainView from '../components/Finance/FinanceMainView';
 import AssetBreakdown from '../components/Finance/AssetBreakdown';
